script.py

+++ script.py
@@ -3,7 +3,10 @@ from imgD import ana
 import os
 
 df = pd.DataFrame([],columns = ['Location','Date','Time','People'])
-df.to_csv('results.csv', mode='a', index = True, header = True)
+if os.path.isfile("results.csv") == False:
+    df.to_csv('results.csv', mode='a', index = True, header = True)
+else:
+    df.to_csv('results.csv', mode='a', index = True, header = False)
 
 while True:
     for img in os.listdir('images'):
@@ -14,6 +17,7 @@ while True:
                 sep = fn.find('T')
                 date = fn[1:sep]
                 time = fn[sep+1:]
+                print("I think there are ",ana(fn)," people in this frame of ",loc," at time ",time)
                 df = df.append({'Location': loc,'Date': date,'Time':time,'People': ana(fn)}, ignore_index=True)
-    df.to_csv('results.csv', mode='a', index = False, header = False)
+    df.to_csv('results.csv', mode='a', index = True, header = False)
     df = pd.DataFrame([],columns = ['Location','Date','Time','People'])
